move pages to app folder and setup new home folder
