--- frontend/pages/universe/index.js
@@ -1,10 +1,0 @@
-import MainContainer from "../../components/UI/MainContainer";
-import ProfileGallery from "../../components/Universe/ProfileGallery";
-
-export default function Universe() {
-  return (
-    <MainContainer>
-      <ProfileGallery />
-    </MainContainer>
-  );
-}
